+++ src/modules/modem/digital/ModemOOK.cpp
@@ -23,7 +23,7 @@ int ModemOOK::checkSampleRate(long long sampleRate, int audioSampleRate) {
     if (sampleRate < 100) {
         return 100;
     }
-    return sampleRate;
+    return (int)sampleRate;
 }
 
 void ModemOOK::demodulate(ModemKit *kit, ModemIQData *input, AudioThreadInput *audioOut) {
